requirements.txt
autogluon[all]
black[jupyter]
botocore
dask[dataframe]
datasets
dvc[all]
dvclive
hydra_colorlog
hydra-core
isort
joblib
numpy
pandas
scikit-learn
seaborn